requirements.txt
langchain
google-genai
langchain-google-genai
requests
python-dotenv 